+++ src/types.ts
@@ -15,7 +15,7 @@ export interface RunInput {
   updateState: UpdateState
   state?: any
   signal: AbortSignal
-  meta?: any
+  context?: any
 }
 
 type Millis = number
@@ -37,12 +37,12 @@ export interface Options {
   includeNodes?: string[]
   excludeNodes?: string[]
   data?: any // Fed into starting nodes (i.e., nodes with no dependencies)
-  meta?: any // Fed to all nodes
+  context?: any // Fed to all nodes
 }
 
 export type Response = {
   emitter: EventEmitter<Events, any>
-  promise: Promise<Snapshot>
+  promise: Promise<void>
   getSnapshot: () => Snapshot
 }
 export type Status = 'pending' | 'running' | 'completed' | 'errored'
@@ -65,9 +65,28 @@ export interface Snapshot {
   finished?: Date
   dag: DAG
   data: SnapshotData
-  meta?: any
+  context?: any
 }
 
 export type ObjectOfPromises = Record<string | number, Promise<any>>
 
 export type Events = 'data' | 'error' | 'done'
+
+export type RunTopology = (
+  spec: Spec,
+  inputDag: DAG,
+  options?: Options
+) => Response
+
+export type RunTopologyInternal = (
+  spec: Spec,
+  snapshot: Snapshot,
+  dag: DAG,
+  context: any
+) => Response
+
+export type ResumeTopology = (
+  spec: Spec,
+  snapshot: Snapshot,
+  options?: Pick<Options, 'context'>
+) => Response