+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.3.0
+
+* `runTopology` and `resumeTopology` return `Promise<void>` from
+the `promise` property.
+* Don't emit `data` when finished.
+* Changed `meta` to `context`. This is just a blob that is passed
+to `runTopology` and `resumeTopology` without needing to be serialized.
+
 # 0.2.0
 
 * Always emit `data`.